chore: move some LabelSet utils to a new package
This is to reduce dependencies on pkg/util, which drags in a whole bunch of other dependencies.

In some cases the replacement is rewritten to cut out an intermediate call.
