+++ pkg/util/conv.go
@@ -2,27 +2,9 @@ package util //nolint:revive
 
 import (
 	"time"
-	"unsafe"
 
 	"github.com/prometheus/common/model"
-	"github.com/prometheus/prometheus/model/labels"
 )
-
-// ModelLabelSetToMap convert a model.LabelSet to a map[string]string
-func ModelLabelSetToMap(m model.LabelSet) map[string]string {
-	if len(m) == 0 {
-		return map[string]string{}
-	}
-	return *(*map[string]string)(unsafe.Pointer(&m)) // #nosec G103 -- we know the string is not mutated -- nosemgrep: use-of-unsafe-block
-}
-
-// MapToModelLabelSet converts a map into a model.LabelSet
-func MapToModelLabelSet(m map[string]string) model.LabelSet {
-	if len(m) == 0 {
-		return model.LabelSet{}
-	}
-	return *(*map[model.LabelName]model.LabelValue)(unsafe.Pointer(&m)) // #nosec G103 -- we know the string is not mutated -- nosemgrep: use-of-unsafe-block
-}
 
 // RoundToMilliseconds returns milliseconds precision time from nanoseconds.
 // from will be rounded down to the nearest milliseconds while through is rounded up.
@@ -36,14 +18,4 @@ func RoundToMilliseconds(from, through time.Time) (model.Time, model.Time) {
 	}
 
 	return model.Time(fromMs), model.Time(throughMs)
-}
-
-// LabelsToMetric converts a Labels to Metric
-// Don't do this on any performance sensitive paths.
-func LabelsToMetric(ls labels.Labels) model.Metric {
-	m := make(model.Metric, ls.Len())
-	ls.Range(func(l labels.Label) {
-		m[model.LabelName(l.Name)] = model.LabelValue(l.Value)
-	})
-	return m
 }
