+++ pkg/util/flagext/labelset.go
@@ -9,7 +9,7 @@ import (
 	"github.com/prometheus/common/model"
 	yaml "go.yaml.in/yaml/v4"
 
-	"github.com/grafana/loki/v3/pkg/util"
+	"github.com/grafana/loki/v3/pkg/util/labels"
 )
 
 // LabelSet is a labelSet that can be used as a flag.
@@ -84,7 +84,7 @@ func (v *LabelSet) UnmarshalYAML(value *yaml.Node) error {
 
 // MarshalYAML implements yaml.Marshaller.
 func (v LabelSet) MarshalYAML() (interface{}, error) {
-	out, err := yaml.Marshal(util.ModelLabelSetToMap(v.LabelSet))
+	out, err := yaml.Marshal(labels.ModelLabelSetToMap(v.LabelSet))
 	if err != nil {
 		return nil, err
 	}
